Homework #3 - Test framework - Доработан класс тестирования, теперь он работает с несколькими тестами в последовательности Before->Test->After для каждого теста. Также рефлекция перенесена в класс стартера.
