+++ HW03-test-framework/build.gradle
@@ -7,4 +7,5 @@ targetCompatibility = JavaVersion.VERSION_17
 
 dependencies {
     compileOnly group: 'org.projectlombok', name: 'lombok', version: '1.18.24'
+    annotationProcessor 'org.projectlombok:lombok:1.18.24'
 }
